requirements.txt
annotated-types==0.7.0
apache-beam==2.63.0
async-timeout==5.0.1
attrs==25.1.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
crcmod==1.7
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
fastavro==1.10.0
fasteners==0.19
Flask==3.1.0
flask-cors==5.0.1
gcloud==0.18.3
google-api-core==2.24.1
google-apitools==0.5.31
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.84.0
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.29.0
google-cloud-bigtable==2.29.0
google-cloud-core==2.4.2
google-cloud-datastore==2.20.2
google-cloud-dlp==3.28.0
google-cloud-language==2.17.0
google-cloud-pubsub==2.28.0
google-cloud-pubsublite==1.12.0
google-cloud-recommendations-ai==0.10.16
google-cloud-resource-manager==1.14.1
google-cloud-spanner==3.52.0
google-cloud-storage==2.19.0
google-cloud-videointelligence==2.16.0
google-cloud-vision==3.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
grpc-google-iam-v1==0.14.1
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.65.5
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
MarkupSafe==3.0.2
more-itertools==10.6.0
numpy==2.2.3
oauth2client==4.1.3
objsize==0.7.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
SecretStorage==3.3.3
shapely==2.0.7
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
